requirements.txt
psutil~=5.9.1
asyncio~=3.4.3

pyrogram>=2.0.51
tgcrypto~=1.2.3
requests~=2.28.1
pymongo[srv]~=4.2.0

wget~=3.2
motor~=3.0.0
pybase64~=1.2.2

ffmpeg-python
py-tgcalls==0.9.7
yt-dlp==2022.10.04

python-dotenv~=0.21.0
aiohttp[speedups]~=3.8.1
httpx[http2]>=0.18.1,<1.0.0
youtube-search-python~=1.6.6
